change fdroid to apache dir

+++ src/main/java/com/m2m/management/controller/RepoAppsController.java
@@ -1,11 +1,14 @@
 package com.m2m.management.controller;
 
 import com.alibaba.fastjson.JSONObject;
+import com.m2m.management.Resource.DeployResource;
+import com.m2m.management.Resource.RepoResource;
 import com.m2m.management.entity.Repo;
 import com.m2m.management.entity.RepoApps;
 import com.m2m.management.repository.IRepoAppsBean;
 
 import com.m2m.management.repository.IRepoBean;
+import com.m2m.management.utils.FileUtil;
 import com.m2m.management.utils.GetApkInfo;
 import com.m2m.management.restful.RepoManager;
 import com.m2m.management.utils.Response;
@@ -32,8 +35,9 @@ import java.util.Optional;
 @Controller
 public class RepoAppsController {
     private static final Logger logger = LoggerFactory.getLogger(RepoAppsController.class);
-    @Value("${apprepo.server}")
-    private String baseUrl;
+    private String baseRepoPath = DeployResource.BASEDEPLOYPATH+ RepoResource.TYPE;
+    private String pathSeparate = File.separator;
+
     @Autowired
     private IRepoAppsBean repoAppsService;
 
@@ -75,7 +79,8 @@ public class RepoAppsController {
             @RequestParam(value = "website", required = false) String website
             ){
         try{
-            File convFile = new File( System.getProperty("java.io.tmpdir")+"/"+file.getOriginalFilename());
+            Boolean isSave = false;
+            File convFile = new File( System.getProperty("java.io.tmpdir")+pathSeparate+file.getOriginalFilename());
             file.transferTo(convFile);
             Map<String, Object> apkInfo =  GetApkInfo.readApk(convFile);
             String filename = file.getOriginalFilename();
@@ -84,16 +89,20 @@ public class RepoAppsController {
             String versionname = apkInfo.get("versionname").toString();
             Optional<Repo> oprepo = repoService.findById(rid);
             Repo repo = oprepo.get();
-            RepoManager repoManager = new RepoManager();
-            ResponseEntity<String> response = repoManager.addFileByReponame(
-                    convFile,
-                    repo.getDarkname(),
-                    description, license,
-                    summary,
-                    website,
-                    filename);
-            logger.info(response.getBody());
-            if(JSONObject.parseObject(response.getBody()).getString("status").equals("success")){
+            String apkSavePath = baseRepoPath + pathSeparate+repo.getDarkname() + pathSeparate + pkgname + pathSeparate + versionname + pathSeparate + filename;
+            logger.info("apkSavePath:"+apkSavePath);
+            File baseDeplyFile = new File(DeployResource.BASEDEPLOYPATH);
+            if(baseDeplyFile.exists()){
+                File apkRepoFile = new File(baseRepoPath + pathSeparate + repo.getDarkname());
+                if(apkRepoFile.exists()){
+                    isSave = FileUtil.copyFile(convFile, apkSavePath);
+                }else{
+                    return new ResponseEntity(Response.error("apk repo is not found"), HttpStatus.NOT_FOUND);
+                }
+            }else{
+                return new ResponseEntity(Response.error("server error"), HttpStatus.NOT_FOUND);
+            }
+            if(isSave){
                 RepoApps repoApps = new RepoApps(filename, pkgname);
                 repoApps.setVersionname(versionname);
                 repoApps.setVersioncode(versioncode);
@@ -107,8 +116,6 @@ public class RepoAppsController {
             }else{
                 return new ResponseEntity(Response.error("add apk error"), HttpStatus.INTERNAL_SERVER_ERROR);
             }
-
-
         }catch (NoSuchElementException e){
             logger.error(e.getMessage());
             return new ResponseEntity(Response.error("rid is error"), HttpStatus.NOT_FOUND);
@@ -142,19 +149,27 @@ public class RepoAppsController {
     @RequestMapping(value = "/repoapps/{rfid}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> deleteRepoAppsById(@PathVariable("rfid") long rfid){
         try{
+            Boolean isDelete = false;
             RepoApps repoApps = repoAppsService.findById(rfid).get();
             Repo repo = repoApps.getRepo();
             String darkname = repo.getDarkname();
             String filename = repoApps.getFilename();
-            RepoManager repoManager = new RepoManager();
-            ResponseEntity<String> response = repoManager.deleteFileByReponame(darkname, filename);
-            repoAppsService.deleteById(rfid);
-            if(JSONObject.parseObject(response.getBody()).getString("status").equals("success")){
+            String pkgname = repoApps.getPkgname();
+            String versionname = repoApps.getVersionname();
+            String pkgPath = baseRepoPath + pathSeparate + darkname + pathSeparate + pkgname;
+            if(FileUtil.isOnlyChildDir(pkgPath)){
+                isDelete = FileUtil.delDir(pkgPath);
+            }else{
+                isDelete = FileUtil.delDir(pkgPath + pathSeparate + versionname);
+            }
+            if(isDelete){
+                repoAppsService.deleteById(rfid);
                 return new ResponseEntity(Response.success(), HttpStatus.OK);
             }else{
                 return new ResponseEntity(Response.error("delete ftp app error"), HttpStatus.NOT_FOUND);
             }
         }catch(NullPointerException e){
+            logger.error(e.getMessage());
             return new ResponseEntity(Response.error("user not found"), HttpStatus.NOT_FOUND);
         }catch(Exception e){
             logger.error(e.getMessage());
@@ -162,14 +177,14 @@ public class RepoAppsController {
         }
     }
 
-    @RequestMapping(value="/repoapps", method = RequestMethod.DELETE)
-    public ResponseEntity<Void> deleteRepoAll(){
-        try{
-            repoAppsService.deleteAll();
-            return new ResponseEntity(Response.success(), HttpStatus.OK);
-        }catch(Exception e){
-            logger.error(e.getMessage());
-            return new ResponseEntity(Response.error("server error"), HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
+//    @RequestMapping(value="/repoapps", method = RequestMethod.DELETE)
+//    public ResponseEntity<Void> deleteRepoAll(){
+//        try{
+//            repoAppsService.deleteAll();
+//            return new ResponseEntity(Response.success(), HttpStatus.OK);
+//        }catch(Exception e){
+//            logger.error(e.getMessage());
+//            return new ResponseEntity(Response.error("server error"), HttpStatus.INTERNAL_SERVER_ERROR);
+//        }
+//    }
 }
